--- CSO_Website/website/members/ccarroll/index.html
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Coming Soon</title>
-    <style>
-        body {
-            font-family: Monospace;
-            background-color: #000000;
-            margin: 0px;
-            overflow: hidden;
-        }
-        #container {
-            height: 100%;
-            width: 100%;
-            display:block;
-            position: relative;
-        }
-        #debug{
-            background-color: #CCCCCC;
-            color:#000000;
-            top:0;
-            padding:10px;
-            width:15%;
-            position: absolute;
-        }
-    </style>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r125/three.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/dat-gui/0.7.7/dat.gui.min.js"></script>
-    <script src="https://threejs.org/examples/js/controls/OrbitControls.js"></script>
-</head>
-<body>
-    <div id="container">
-    </div>
-    <p id="debug">
-    </p>
-    <script src="comingsoon.js">
-    </script>
-</body>
-</html>
