--- CSO_Website/website/members/jwinward/index.html
@@ -1,8 +1,0 @@
-<html>
-<body>
-
-<p><b>Hi Female Parent! Hi Male Parent! Hi Sibling!</b></p>
-<p>-John Winward</p>
-
-</body>
-</html>
