--- CSO_Website/website/members/smcmanus/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>Shawn McManus</title>
-</head>
-<body>
-
-<h1>:O</h1>
-
-</body>
-</html>
