--- CSO_Website/website/static/css/master.css
@@ -1,138 +1,0 @@
-/*
-*****************************
-*****************************
-*** FONT ********************
-*****************************
-*****************************
-*/
-
-@font-face {
-    font-family: logoFont;
-    src: url(logofont.ttf);
-}
-
-
-/*
-*****************************
-*****************************
-*** COMING SOON *************
-*****************************
-*****************************
-*/
-
-body {
-    background-color: #32353b;
-}
-
-.console-container {
-  font-family: logoFont;
-  font-size: 3em;
-  text-align: center;
-  height: 200px;
-  width: 600px;
-  display: block;
-  position: relative;
-  color: white;
-  top: 100px;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  margin: auto;
-}
-
-.console-underscore {
-  display: inline-block;
-  position: relative;
-  top: 0;
-  left: -30px;
-}
-
-.front-pic {
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  border: solid 2px black;
-  border-radius: 25px;
-}
-
-
-/*
-*****************************
-*****************************
-*** LOGO BACKGROUND *********
-*****************************
-*****************************
-*/
-
-.logo-background { position: relative; }
-
-.logo-background .orange-red-crayola, .flame, .honey-yellow,
-                 .mountain-meadow, .sea-green-crayola, .carolina-blue {
-  width: 100%;
-  height: 20px;
-  border-style: solid;
-  border-width: 2px;
-  border-color: #161613;
-  border-radius: 5px;
-}
-
-.logo-background .orange-red-crayola { background-color: #FE5F55; }
-
-.logo-background .flame { background-color: #E4572E; }
-
-.logo-background .honey-yellow { background-color: #F7B32B; }
-
-.logo-background .mountain-meadow { background-color: #13C4A3; }
-
-.logo-background .sea-green-crayola { background-color: #36F1CD; }
-
-.logo-background .carolina-blue { background-color: #39A0ED; }
-
-.logo-background .logo-image {
-  height: 65%;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-
-/*
-*****************************
-*****************************
-*** NAV MENU ****************
-*****************************
-*****************************
-*/
-
-.navbar {
-  margin-bottom: 0;
-}
-
-.below-nav {
-  width: 100%;
-  height: 20px;
-  border-style: solid;
-  border-width: 2px;
-  border-color: #161613;
-  border-radius: 5px;
-  background-color: #39A0ED;
-}
-
-.navbar-nav.navbar-center li {
-  font-family: logoFont;
-  font-size: 200%;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 5px;
-}
-
-.navbar .navbar-nav {
-  display: inline-block;
-  float: none;
-  vertical-align: top;
-}
-
-.navbar .navbar-collapse {
-  text-align: center;
-}
