--- CSO_Website/website/templates/index.html
@@ -1,69 +1,0 @@
-﻿<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <title>CSO_CRINGE</title>
-
-    <!-- BOOTSTRAP -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
-
-    <!-- CSS -->
-    <link rel="stylesheet" href="assets/css/master.css">
-
-    <!-- GOOGLE FONTS -->
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Mukta:wght@600&display=swap" rel="stylesheet">
-  </head>
-  <body>
-
-    <section id="logo-top">
-      <div class="logo-background">
-        <div class="orange-red-crayola"></div>
-        <div class="flame"></div>
-        <div class="honey-yellow"></div>
-        <div class="mountain-meadow"></div>
-        <div class="sea-green-crayola"></div>
-        <div class="carolina-blue"></div>
-        <div class="logo-image-container">
-          <img class="logo-image" src="assets/images/CSO_BLEK.png" alt="Missing Image">
-        </div>
-      </div>
-    </section>
-
-    <section id="navigation-bar">
-      <nav class="navbar navbar-inverse navbar-expand-lg navbar-static-top">
-        <div class="container-fluid">
-          <button class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar-elements">
-              <span class="icon-bar"></span>
-              <span class="icon-bar"></span>
-              <span class="icon-bar"></span>
-          </button>
-          <div class="collapse navbar-collapse justify-content-center" id="navbar-elements">
-              <ul class="nav navbar-nav navbar-center">
-                  <li class="nav-item"><a href="#">Home</a></li>
-                  <li class="nav-item"><a href="https://esucso.org/members/members-index.html">Members</a></li>
-                  <li class="nav-item"><a href="#">Events</a></li>
-                  <li class="nav-item"><a href="#">About Us</a></li>
-              </ul>
-          </div>
-        </div>
-      </nav>
-      <div class="below-nav"></div>
-    </section>
-
-    <div class='console-container'>
-      <span id='text'></span>
-      <div class='console-underscore' id='console'></div>
-    </div>
-
-    <div class="container">
-      <img class="front-pic" src="assets/images/cat-keyboard.gif" alt="">
-    </div>
-
-    <!-- Javascript -->
-    <script   src="https://code.jquery.com/jquery-3.1.1.min.js"        integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8="     crossorigin="anonymous"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
-    <script src="assets/js/console_text.js"></script>
-
-  </body>
-</html>
