+++ website_project/templates/base.html
@@ -1,0 +1,61 @@
+<!DOCTYPE html>
+{% load static %}
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>CSO</title>
+
+    <!-- BOOTSTRAP -->
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+
+    <!--  CSS  -->
+    <link rel="stylesheet" href="{% static 'css/master.css' %}">
+</head>
+<body>
+
+    <section id="logo-top">
+      <div class="logo-background">
+        <div class="orange-red-crayola"></div>
+        <div class="flame"></div>
+        <div class="honey-yellow"></div>
+        <div class="mountain-meadow"></div>
+        <div class="sea-green-crayola"></div>
+        <div class="carolina-blue"></div>
+        <div class="logo-image-container">
+          <img class="logo-image" src="{% static 'images/CSO_BLEK.png' %}" alt="Missing Image">
+        </div>
+      </div>
+    </section>
+
+    <section id="navigation-bar">
+      <nav class="navbar navbar-inverse navbar-expand-lg navbar-static-top">
+        <div class="container-fluid">
+          <button class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar-elements">
+              <span class="icon-bar"></span>
+              <span class="icon-bar"></span>
+              <span class="icon-bar"></span>
+          </button>
+          <div class="collapse navbar-collapse justify-content-center" id="navbar-elements">
+              <ul class="nav navbar-nav navbar-center">
+                  <li class="nav-item"><a href="#">Home</a></li>
+                  <li class="nav-item"><a href="https://esucso.org/members/members-index.html">Members</a></li>
+                  <li class="nav-item"><a href="#">Events</a></li>
+                  <li class="nav-item"><a href="#">About Us</a></li>
+              </ul>
+          </div>
+        </div>
+      </nav>
+      <div class="below-nav"></div>
+    </section>
+
+    {% block content %}
+
+    {% endblock %}
+
+    <!-- Javascript -->
+    <script   src="https://code.jquery.com/jquery-3.1.1.min.js"        integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8="     crossorigin="anonymous"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
+    <script src="{% static 'js/console_text.js' %}"></script>
+
+</body>
+</html>
