+++ website_project/templates/index.html
@@ -1,0 +1,12 @@
+{% extends 'base.html' %}
+{% load static %}
+{% block content %}
+     <div class='console-container'>
+      <span id='text'></span>
+      <div class='console-underscore' id='console'></div>
+    </div>
+
+    <div class="container">
+      <img class="front-pic" src="{% static 'images/cat-keyboard.gif' %}" alt="">
+    </div>
+{% endblock %}
