+++ website_project/website_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class WebsiteAppConfig(AppConfig):
+    name = 'website_app'
