+++ website_project/website_app/views.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+from django.views.generic import TemplateView
+
+
+# Create your views here.
+class HomePage(TemplateView):
+    template_name = 'index.html'
